refactor: Move the Stream traits into the stream module
BREAKING CHANGE

The `Stream*` traits (and `Positioned`) are now in the stream module. Change any references to them through `primitives` to `stream`
