+++ src/primitives.rs
@@ -1,8 +1,6 @@
 use lib::fmt;
 
 #[cfg(feature = "std")]
-use std::io::{Bytes, Read};
-#[cfg(feature = "std")]
 use std::error::Error as StdError;
 
 use either::Either;
@@ -13,8 +11,7 @@ use ErrorOffset;
 use combinator::{and_then, expected, flat_map, map, message, or, skip, then, then_partial, with,
                  AndThen, Expected, FlatMap, Iter, Map, Message, Or, Skip, Then, ThenPartial, With};
 
-#[cfg(feature = "std")]
-use easy::Error;
+use stream::{Resetable, Stream, StreamOnce};
 
 #[macro_export]
 macro_rules! ctry {
@@ -32,36 +29,11 @@ macro_rules! ctry {
     }
 }
 
-/// Newtype around a pointer offset into a slice stream (`&[T]`/`&str`).
-#[derive(Clone, Copy, Debug, Default, Eq, PartialEq, Ord, PartialOrd)]
-pub struct PointerOffset(pub usize);
-
-impl fmt::Display for PointerOffset {
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        write!(f, "{:?}", self.0 as *const ())
-    }
-}
-
-impl PointerOffset {
-    /// Converts the pointer-based position into an indexed position.
-    ///
-    /// ```rust
-    /// # extern crate combine;
-    /// # use combine::*;
-    /// # fn main() {
-    /// let text = "b";
-    /// let err = token('a').easy_parse(text).unwrap_err();
-    /// assert_eq!(err.position.0, text.as_ptr() as usize);
-    /// assert_eq!(err.map_position(|p| p.translate_position(text)).position, 0);
-    /// # }
-    /// ```
-    pub fn translate_position<T>(mut self, initial_string: &T) -> usize
-    where
-        T: ?Sized,
-    {
-        self.0 -= initial_string as *const T as *const () as usize;
-        self.0
-    }
+#[derive(Clone, Debug)]
+pub enum Info<T, R> {
+    Token(T),
+    Range(R),
+    Borrowed(&'static str),
 }
 
 impl<R> From<char> for Info<char, R> {
@@ -232,24 +204,6 @@ impl<T> Consumed<T> {
 /// `I` is the specific stream type used in the parser.
 pub type ParseResult<O, I> = Result<(O, Consumed<()>), Consumed<Tracked<<I as StreamOnce>::Error>>>;
 pub type ParseResult2<O, E> = Result<(O, Consumed<()>), Consumed<Tracked<E>>>;
-
-#[derive(Clone, Debug)]
-pub enum Info<T, R> {
-    Token(T),
-    Range(R),
-    Borrowed(&'static str),
-}
-
-/// Enum used to store information about an error that has occurred during parsing.
-#[derive(Debug)]
-pub enum EasyError<T, R> {
-    /// Error indicating an unexpected token has been encountered in the stream
-    Unexpected(Info<T, R>),
-    /// Error indicating that the parser expected something else
-    Expected(Info<T, R>),
-    /// Generic message
-    Message(Info<T, R>),
-}
 
 /// `StreamError` represents a single error returned from a `Stream` or a `Parser`.
 /// Usually this is composed into a `ParseError`
@@ -505,337 +459,6 @@ where
     }
 }
 
-pub trait Positioned: StreamOnce {
-    /// Returns the current position of the stream.
-    fn position(&self) -> Self::Position;
-}
-
-/// `StreamOnce` represents a sequence of items that can be extracted one by one.
-pub trait StreamOnce {
-    /// The type of items which is yielded from this stream.
-    type Item: Clone + PartialEq;
-
-    /// The type of a range of items yielded from this stream.
-    /// Types which do not a have a way of yielding ranges of items should just use the
-    /// `Self::Item` for this type.
-    type Range: Clone + PartialEq;
-
-    /// Type which represents the position in a stream.
-    /// `Ord` is required to allow parsers to determine which of two positions are further ahead.
-    type Position: Clone + Ord;
-
-    type Error: ParseError<Self::Item, Self::Range, Self::Position>;
-    /// Takes a stream and removes its first item, yielding the item and the rest of the elements.
-    /// Returns `Err` if no element could be retrieved.
-
-    fn uncons<E>(&mut self) -> Result<Self::Item, E>
-    where
-        E: StreamError<Self::Item, Self::Range>;
-
-    fn is_partial(&self) -> bool {
-        false
-    }
-}
-
-pub trait Resetable {
-    type Checkpoint: Clone;
-
-    fn checkpoint(&self) -> Self::Checkpoint;
-    fn reset(&mut self, checkpoint: Self::Checkpoint);
-}
-
-#[macro_export]
-macro_rules! clone_resetable {
-    (( $($params: tt)* ) $ty: ty) => {
-        impl<$($params)*> Resetable for $ty
-        {
-            type Checkpoint = Self;
-
-            fn checkpoint(&self) -> Self {
-                self.clone()
-            }
-            fn reset(&mut self, checkpoint: Self) {
-                *self = checkpoint;
-            }
-        }
-    }
-}
-
-clone_resetable!{('a) &'a str}
-clone_resetable!{('a, T) &'a [T]}
-clone_resetable!{('a, T) SliceStream<'a, T> }
-clone_resetable!{(T: Clone) IteratorStream<T>}
-
-/// A stream of tokens which can be duplicated
-pub trait Stream: StreamOnce + Resetable + Positioned {}
-
-impl<I> Stream for I
-where
-    I: StreamOnce + Positioned + Resetable,
-    I::Error: ParseError<I::Item, I::Range, I::Position>,
-{
-}
-
-#[inline]
-pub fn uncons<I>(input: &mut I) -> ConsumedResult<I::Item, I>
-where
-    I: ?Sized + Stream,
-{
-    match input.uncons() {
-        Ok(x) => ConsumedOk(x),
-        Err(err) => wrap_stream_error(input, err),
-    }
-}
-
-/// A `RangeStream` is an extension of `StreamOnce` which allows for zero copy parsing.
-pub trait RangeStreamOnce: StreamOnce + Resetable {
-    /// Takes `size` elements from the stream.
-    /// Fails if the length of the stream is less than `size`.
-    fn uncons_range<E>(&mut self, size: usize) -> Result<Self::Range, E>
-    where
-        E: StreamError<Self::Item, Self::Range>;
-
-    /// Takes items from stream, testing each one with `predicate`.
-    /// returns the range of items which passed `predicate`.
-    fn uncons_while<E, F>(&mut self, f: F) -> Result<Self::Range, E>
-    where
-        F: FnMut(Self::Item) -> bool,
-        E: StreamError<Self::Item, Self::Range>;
-
-    /// Returns the distance between `self` and `end`. The returned `usize` must be so that
-    ///
-    /// ```ignore
-    /// let start = stream.checkpoint();
-    /// stream.uncons_range(distance);
-    /// stream.distance(&start) == distance
-    /// ```
-    fn distance(&self, end: &Self::Checkpoint) -> usize;
-}
-
-/// A `RangeStream` is an extension of `Stream` which allows for zero copy parsing.
-pub trait RangeStream: Stream + RangeStreamOnce {}
-
-impl<I> RangeStream for I
-where
-    I: RangeStreamOnce + Stream,
-{
-}
-
-/// A `RangeStream` which is capable of providing it's entire range.
-pub trait FullRangeStream: RangeStream {
-    /// Returns the entire range of `self`
-    fn range(&self) -> Self::Range;
-}
-
-#[doc(hidden)]
-#[inline]
-pub fn wrap_stream_error<T, I>(
-    input: &I,
-    err: <I::Error as ParseError<I::Item, I::Range, I::Position>>::StreamError,
-) -> ConsumedResult<T, I>
-where
-    I: ?Sized + StreamOnce + Positioned,
-{
-    let err = I::Error::from_error(input.position(), err);
-    if input.is_partial() {
-        ConsumedErr(err)
-    } else {
-        EmptyErr(err.into())
-    }
-}
-
-#[inline]
-pub fn uncons_range<I>(input: &mut I, size: usize) -> ConsumedResult<I::Range, I>
-where
-    I: ?Sized + RangeStream,
-    I::Range: Range,
-{
-    match input.uncons_range(size) {
-        Err(err) => wrap_stream_error(input, err),
-        Ok(x) => if x.len() == 0 {
-            EmptyOk(x)
-        } else {
-            ConsumedOk(x)
-        },
-    }
-}
-
-fn input_at_eof<I>(input: &mut I) -> bool
-where
-    I: ?Sized + Stream,
-{
-    let before = input.checkpoint();
-    let x = input.uncons::<UnexpectedParse>() == Err(UnexpectedParse::Eoi);
-    input.reset(before);
-    x
-}
-
-/// Removes items from the input while `predicate` returns `true`.
-#[inline]
-pub fn uncons_while<I, F>(input: &mut I, predicate: F) -> ConsumedResult<I::Range, I>
-where
-    F: FnMut(I::Item) -> bool,
-    I: ?Sized + RangeStream,
-    I::Range: Range,
-{
-    match input.uncons_while(predicate) {
-        Err(err) => wrap_stream_error(input, err),
-        Ok(x) => {
-            if input.is_partial() && input_at_eof(input) {
-                // Partial inputs which encounter end of file must fail to let more input be
-                // retrieved
-                ConsumedErr(I::Error::from_error(
-                    input.position(),
-                    StreamError::end_of_input(),
-                ))
-            } else {
-                if x.len() == 0 {
-                    EmptyOk(x)
-                } else {
-                    ConsumedOk(x)
-                }
-            }
-        }
-    }
-}
-
-pub trait Range {
-    /// Returns the remaining length of `self`.
-    /// The returned length need not be the same as the number of items left in the stream.
-    fn len(&self) -> usize;
-
-    /// Returns `true` if the range does not contain any elements (`Range::len() == 0`)
-    fn is_empty(&self) -> bool {
-        self.len() == 0
-    }
-}
-
-impl<'a> RangeStreamOnce for &'a str {
-    #[inline]
-    fn uncons_while<E, F>(&mut self, mut f: F) -> Result<&'a str, E>
-    where
-        F: FnMut(Self::Item) -> bool,
-        E: StreamError<Self::Item, Self::Range>,
-    {
-        let mut chars = self.chars();
-        while let Some(c) = chars.next() {
-            if !f(c) {
-                let len = self.len() - chars.as_str().len() - c.len_utf8();
-                let (result, rest) = self.split_at(len);
-                *self = rest;
-                return Ok(result);
-            }
-        }
-        let result = *self;
-        *self = &self[self.len()..];
-        Ok(result)
-    }
-
-    #[inline]
-    fn uncons_range<E>(&mut self, size: usize) -> Result<&'a str, E>
-    where
-        E: StreamError<Self::Item, Self::Range>,
-    {
-        fn is_char_boundary(s: &str, index: usize) -> bool {
-            if index == s.len() {
-                return true;
-            }
-            match s.as_bytes().get(index) {
-                None => false,
-                Some(&b) => b < 128 || b >= 192,
-            }
-        }
-        if size <= self.len() {
-            if is_char_boundary(self, size) {
-                let (result, remaining) = self.split_at(size);
-                *self = remaining;
-                Ok(result)
-            } else {
-                Err(E::message_static_message(CHAR_BOUNDARY_ERROR_MESSAGE))
-            }
-        } else {
-            Err(E::end_of_input())
-        }
-    }
-
-    #[inline]
-    fn distance(&self, end: &Self) -> usize {
-        self.position().0 - end.position().0
-    }
-}
-
-impl<'a> FullRangeStream for &'a str {
-    fn range(&self) -> Self::Range {
-        self
-    }
-}
-
-impl<'a> Range for &'a str {
-    #[inline]
-    fn len(&self) -> usize {
-        str::len(self)
-    }
-}
-
-impl<'a, T> Range for &'a [T] {
-    #[inline]
-    fn len(&self) -> usize {
-        <[T]>::len(self)
-    }
-}
-
-impl<'a, T> RangeStreamOnce for &'a [T]
-where
-    T: Clone + PartialEq,
-{
-    #[inline]
-    fn uncons_range<E>(&mut self, size: usize) -> Result<&'a [T], E>
-    where
-        E: StreamError<Self::Item, Self::Range>,
-    {
-        if size <= self.len() {
-            let (result, remaining) = self.split_at(size);
-            *self = remaining;
-            Ok(result)
-        } else {
-            Err(E::end_of_input())
-        }
-    }
-
-    #[inline]
-    fn uncons_while<E, F>(&mut self, mut f: F) -> Result<&'a [T], E>
-    where
-        F: FnMut(Self::Item) -> bool,
-        E: StreamError<Self::Item, Self::Range>,
-    {
-        let len = self.iter().take_while(|c| f((**c).clone())).count();
-        let (result, remaining) = self.split_at(len);
-        *self = remaining;
-        Ok(result)
-    }
-
-    #[inline]
-    fn distance(&self, end: &Self) -> usize {
-        end.len() - self.len()
-    }
-}
-
-impl<'a, T> FullRangeStream for &'a [T]
-where
-    T: Clone + PartialEq,
-{
-    fn range(&self) -> Self::Range {
-        self
-    }
-}
-
-impl<'a> Positioned for &'a str {
-    #[inline(always)]
-    fn position(&self) -> Self::Position {
-        self.as_bytes().position()
-    }
-}
-
 #[derive(Clone, Copy, Debug, PartialEq)]
 pub enum StringStreamError {
     UnexpectedParse,
@@ -843,7 +466,7 @@ pub enum StringStreamError {
     CharacterBoundary,
 }
 
-const CHAR_BOUNDARY_ERROR_MESSAGE: &str = "unexpected slice on character boundary";
+pub(crate) const CHAR_BOUNDARY_ERROR_MESSAGE: &str = "unexpected slice on character boundary";
 
 impl fmt::Display for StringStreamError {
     fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
@@ -972,353 +595,6 @@ where
         T: ParseError<Item, Range, Position>,
     {
         T::from_error(Position::default(), StreamError::into_other(self))
-    }
-}
-
-impl<'a> StreamOnce for &'a str {
-    type Item = char;
-    type Range = &'a str;
-    type Position = PointerOffset;
-    type Error = StringStreamError;
-
-    #[inline]
-    fn uncons<E>(&mut self) -> Result<char, E>
-    where
-        E: StreamError<Self::Item, Self::Range>,
-    {
-        let mut chars = self.chars();
-        match chars.next() {
-            Some(c) => {
-                *self = chars.as_str();
-                Ok(c)
-            }
-            None => Err(E::end_of_input()),
-        }
-    }
-}
-
-impl<'a, T> Positioned for &'a [T]
-where
-    T: Clone + PartialEq,
-{
-    #[inline(always)]
-    fn position(&self) -> Self::Position {
-        PointerOffset(self.as_ptr() as usize)
-    }
-}
-
-impl<'a, T> StreamOnce for &'a [T]
-where
-    T: Clone + PartialEq,
-{
-    type Item = T;
-    type Range = &'a [T];
-    type Position = PointerOffset;
-    type Error = UnexpectedParse;
-
-    #[inline]
-    fn uncons<E>(&mut self) -> Result<T, E>
-    where
-        E: StreamError<Self::Item, Self::Range>,
-    {
-        match self.split_first() {
-            Some((first, rest)) => {
-                *self = rest;
-                Ok(first.clone())
-            }
-            None => Err(E::end_of_input()),
-        }
-    }
-}
-
-/// Stream type which indicates that the stream is partial if end of input is reached
-#[derive(Copy, Clone, Eq, PartialEq, Ord, PartialOrd, Debug)]
-pub struct PartialStream<S>(pub S);
-
-impl<S> Positioned for PartialStream<S>
-where
-    S: Positioned,
-{
-    #[inline(always)]
-    fn position(&self) -> Self::Position {
-        self.0.position()
-    }
-}
-
-impl<S> Resetable for PartialStream<S>
-where
-    S: Resetable,
-{
-    type Checkpoint = S::Checkpoint;
-
-    #[inline(always)]
-    fn checkpoint(&self) -> Self::Checkpoint {
-        self.0.checkpoint()
-    }
-
-    #[inline(always)]
-    fn reset(&mut self, checkpoint: Self::Checkpoint) {
-        self.0.reset(checkpoint);
-    }
-}
-
-impl<S> StreamOnce for PartialStream<S>
-where
-    S: StreamOnce,
-{
-    type Item = S::Item;
-    type Range = S::Range;
-    type Position = S::Position;
-    type Error = S::Error;
-
-    #[inline(always)]
-    fn uncons<E>(&mut self) -> Result<S::Item, E>
-    where
-        E: StreamError<Self::Item, Self::Range>,
-    {
-        self.0.uncons()
-    }
-
-    fn is_partial(&self) -> bool {
-        true
-    }
-}
-
-impl<S> RangeStreamOnce for PartialStream<S>
-where
-    S: RangeStreamOnce,
-{
-    #[inline(always)]
-    fn uncons_range<E>(&mut self, size: usize) -> Result<Self::Range, E>
-    where
-        E: StreamError<Self::Item, Self::Range>,
-    {
-        self.0.uncons_range(size)
-    }
-
-    #[inline(always)]
-    fn uncons_while<E, F>(&mut self, f: F) -> Result<Self::Range, E>
-    where
-        F: FnMut(Self::Item) -> bool,
-        E: StreamError<Self::Item, Self::Range>,
-    {
-        self.0.uncons_while(f)
-    }
-
-    #[inline(always)]
-    fn distance(&self, end: &Self::Checkpoint) -> usize {
-        self.0.distance(end)
-    }
-}
-
-impl<S> FullRangeStream for PartialStream<S>
-where
-    S: FullRangeStream,
-{
-    #[inline(always)]
-    fn range(&self) -> Self::Range {
-        self.0.range()
-    }
-}
-
-/// Newtype for constructing a stream from a slice where the items in the slice are not copyable.
-#[derive(Copy, Eq, PartialEq, Ord, PartialOrd, Debug)]
-pub struct SliceStream<'a, T: 'a>(pub &'a [T]);
-
-impl<'a, T> Clone for SliceStream<'a, T> {
-    fn clone(&self) -> SliceStream<'a, T> {
-        SliceStream(self.0)
-    }
-}
-
-impl<'a, T> Positioned for SliceStream<'a, T>
-where
-    T: PartialEq + 'a,
-{
-    #[inline(always)]
-    fn position(&self) -> Self::Position {
-        PointerOffset(self.0.as_ptr() as usize)
-    }
-}
-
-impl<'a, T> StreamOnce for SliceStream<'a, T>
-where
-    T: PartialEq + 'a,
-{
-    type Item = &'a T;
-    type Range = &'a [T];
-    type Position = PointerOffset;
-    type Error = UnexpectedParse;
-
-    #[inline]
-    fn uncons<E>(&mut self) -> Result<&'a T, E>
-    where
-        E: StreamError<Self::Item, Self::Range>,
-    {
-        match self.0.split_first() {
-            Some((first, rest)) => {
-                self.0 = rest;
-                Ok(first)
-            }
-            None => Err(E::end_of_input()),
-        }
-    }
-}
-
-impl<'a, T> RangeStreamOnce for SliceStream<'a, T>
-where
-    T: PartialEq + 'a,
-{
-    #[inline]
-    fn uncons_range<E>(&mut self, size: usize) -> Result<&'a [T], E>
-    where
-        E: StreamError<Self::Item, Self::Range>,
-    {
-        if size <= self.0.len() {
-            let (range, rest) = self.0.split_at(size);
-            self.0 = rest;
-            Ok(range)
-        } else {
-            Err(E::end_of_input())
-        }
-    }
-
-    #[inline]
-    fn uncons_while<E, F>(&mut self, mut f: F) -> Result<&'a [T], E>
-    where
-        F: FnMut(Self::Item) -> bool,
-        E: StreamError<Self::Item, Self::Range>,
-    {
-        let len = self.0.iter().take_while(|c| f(*c)).count();
-        let (range, rest) = self.0.split_at(len);
-        self.0 = rest;
-        Ok(range)
-    }
-
-    #[inline]
-    fn distance(&self, end: &Self) -> usize {
-        end.0.len() - self.0.len()
-    }
-}
-
-impl<'a, T> FullRangeStream for SliceStream<'a, T>
-where
-    T: PartialEq + 'a,
-{
-    fn range(&self) -> Self::Range {
-        self.0
-    }
-}
-
-/// Wrapper around iterators which allows them to be treated as a stream.
-/// Returned by [`from_iter`].
-///
-/// [`from_iter`]: fn.from_iter.html
-#[derive(Copy, Clone, Debug)]
-pub struct IteratorStream<I>(I);
-
-impl<I> IteratorStream<I>
-where
-    I: Iterator,
-{
-    /// Converts an `Iterator` into a stream.
-    ///
-    /// NOTE: This type do not implement `Positioned` and `Clone` and must be wrapped with types
-    ///     such as `BufferedStreamRef` and `State` to become a `Stream` which can be parsed
-    pub fn new(iter: I) -> IteratorStream<I> {
-        IteratorStream(iter)
-    }
-}
-
-impl<I> Iterator for IteratorStream<I>
-where
-    I: Iterator,
-{
-    type Item = I::Item;
-    fn next(&mut self) -> Option<I::Item> {
-        self.0.next()
-    }
-}
-
-impl<I: Iterator> StreamOnce for IteratorStream<I>
-where
-    I::Item: Clone + PartialEq,
-{
-    type Item = I::Item;
-    type Range = I::Item;
-    type Position = ();
-    type Error = UnexpectedParse;
-
-    #[inline]
-    fn uncons<E>(&mut self) -> Result<I::Item, E>
-    where
-        E: StreamError<Self::Item, Self::Range>,
-    {
-        match self.next() {
-            Some(x) => Ok(x),
-            None => Err(E::end_of_input()),
-        }
-    }
-}
-
-#[cfg(feature = "std")]
-pub struct ReadStream<R> {
-    bytes: Bytes<R>,
-}
-
-#[cfg(feature = "std")]
-impl<R: Read> StreamOnce for ReadStream<R> {
-    type Item = u8;
-    type Range = u8;
-    type Position = ();
-    type Error = Error<u8, u8>;
-
-    #[inline]
-    fn uncons<E>(&mut self) -> Result<u8, E>
-    where
-        E: StreamError<Self::Item, Self::Range>,
-    {
-        match self.bytes.next() {
-            Some(Ok(b)) => Ok(b),
-            Some(Err(err)) => Err(E::other(err)),
-            None => Err(E::end_of_input()),
-        }
-    }
-}
-
-#[cfg(feature = "std")]
-impl<R> ReadStream<R>
-where
-    R: Read,
-{
-    /// Creates a `StreamOnce` instance from a value implementing `std::io::Read`.
-    ///
-    /// NOTE: This type do not implement `Positioned` and `Clone` and must be wrapped with types
-    ///     such as `BufferedStreamRef` and `State` to become a `Stream` which can be parsed
-    ///
-    /// ```rust
-    /// # #![cfg(feature = "std")]
-    /// # extern crate combine;
-    /// use combine::*;
-    /// use combine::byte::*;
-    /// use combine::primitives::ReadStream;
-    /// use combine::buffered_stream::BufferedStream;
-    /// use combine::state::State;
-    /// use std::io::Read;
-    ///
-    /// # fn main() {
-    /// let input: &[u8] = b"123,";
-    /// let stream = BufferedStream::new(State::new(ReadStream::new(input)), 1);
-    /// let result = (many(digit()), byte(b','))
-    ///     .parse(stream)
-    ///     .map(|t| t.0);
-    /// assert_eq!(result, Ok((vec![b'1', b'2', b'3'], b',')));
-    /// # }
-    /// ```
-    pub fn new(read: R) -> ReadStream<R> {
-        ReadStream {
-            bytes: read.bytes(),
-        }
     }
 }
 
@@ -2397,40 +1673,6 @@ where
     #[inline(always)]
     fn add_error(&mut self, error: &mut Tracked<<Self::Input as StreamOnce>::Error>) {
         (**self).add_error(error)
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    #[test]
-    #[inline]
-    fn uncons_range_at_end() {
-        assert_eq!("".uncons_range::<UnexpectedParse>(0), Ok(""));
-        assert_eq!("123".uncons_range::<UnexpectedParse>(3), Ok("123"));
-        assert_eq!((&[1][..]).uncons_range::<UnexpectedParse>(1), Ok(&[1][..]));
-        let s: &[u8] = &[];
-        assert_eq!(
-            SliceStream(s).uncons_range::<UnexpectedParse>(0),
-            Ok(&[][..])
-        );
-    }
-
-    #[test]
-    fn larger_than_1_byte_items_return_correct_distance() {
-        let mut input = &[123i32, 0i32][..];
-
-        let before = input.checkpoint();
-        assert_eq!(input.distance(&before), 0);
-
-        input.uncons::<UnexpectedParse>().unwrap();
-        assert_eq!(input.distance(&before), 1);
-
-        input.uncons::<UnexpectedParse>().unwrap();
-        assert_eq!(input.distance(&before), 2);
-
-        input.reset(before.clone());
-        assert_eq!(input.distance(&before), 0);
     }
 }
